Neo Encoder setup

+++ src/main/java/ca/team4519/frc2020/Gains.java
@@ -3,7 +3,7 @@ package ca.team4519.frc2020;
 public class Gains{
 
 	public static double CONTROL_LOOP_TIME_SECONDS = 0.005;	//200hz
-	public static double CONTROL_LOOP_TIME_MILLISECOND = 5;	//Neos require this conversion
+	public static int CONTROL_LOOP_TIME_MILLISECOND = 5;	//Neos require this conversion
 	public static double NEO_TicksPerRev = 42.0;
 
 	public static final class Drive {
